+++ src/App/Http/Controllers/UserController.php
@@ -1,8 +1,13 @@
 <?php namespace Redooor\Redminportal\App\Http\Controllers;
 
-use Lang;
 use Auth;
+use Hash;
+use Input;
+use Lang;
+use Validator;
+use Illuminate\Support\MessageBag;
 use Redooor\Redminportal\App\Http\Traits\SorterController;
+use Redooor\Redminportal\App\Http\Traits\PermissibleController;
 use Redooor\Redminportal\App\Models\User;
 use Redooor\Redminportal\App\Models\Group;
 
@@ -13,7 +18,7 @@ class UserController extends Controller
     protected $sortBy;
     protected $orderBy;
     
-    use SorterController;
+    use SorterController, PermissibleController;
     
     public function __construct(User $model)
     {
@@ -56,7 +61,7 @@ class UserController extends Controller
         $user = User::find($sid);
         
         if ($user == null) {
-            $errors = new \Illuminate\Support\MessageBag;
+            $errors = new MessageBag;
             $errors->add(
                 'editError',
                 Lang::get('redminportal::messages.user_error_user_not_found')
@@ -71,10 +76,29 @@ class UserController extends Controller
             $groups[$group->id] = $group->id;
         }
         
+        $permission_inherit = [];
+        $permission_allow = [];
+        $permission_deny = [];
+        
+//        if ($user->permissions()) {
+            foreach ($user->permissions() as $key => $value) {
+                if ($value < 0) {
+                    $permission_deny[$key] = $key;
+                } elseif ($value > 0) {
+                    $permission_allow[$key] = $key;
+                } else {
+                    $permission_inherit[$key] = $key;
+                }
+            }
+//        }
+        
         $data = array(
             'roles' => $roles,
             'user' => $user,
-            'groups' => $groups
+            'groups' => $groups,
+            'permission_inherit' => implode(',', $permission_inherit),
+            'permission_allow' => implode(',', $permission_allow),
+            'permission_deny' => implode(',', $permission_deny)
         );
         
         return view('redminportal::users/edit', $data);
@@ -82,53 +106,55 @@ class UserController extends Controller
 
     public function postStore()
     {
-        $sid = \Input::get('id');
+        $sid = Input::get('id');
+        $errors = new MessageBag;
         
         $rules = array(
             'first_name'    => 'required',
             'last_name'     => 'required',
             'role'          => 'required',
-            'email'         => 'required|unique:users,email,' . $sid
+            'email'         => 'required|unique:users,email,' . $sid,
+            'permission-inherit' => 'regex:/^[a-z,0-9._\-?]+$/i',
+            'permission-allow'   => 'regex:/^[a-z,0-9._\-?]+$/i',
+            'permission-deny'    => 'regex:/^[a-z,0-9._\-?]+$/i'
         );
+        
+        // Get activated input first, for checking if user is deactivating own account
+        $activated = (Input::get('activated') == '' ? false : true);
         
         if (isset($sid)) {
             $rules['password'] = 'confirmed|min:6';
-        } else {
-            $rules['password'] = 'required|confirmed|min:6';
-        }
-
-        $validation = \Validator::make(\Input::all(), $rules);
-        
-        $path = (isset($sid) ? 'admin/users/edit/' . $sid : 'admin/users/create');
-        
-        if ($validation->fails()) {
-            return redirect($path)->withErrors($validation)->withInput();
-        }
-
-        $first_name    = \Input::get('first_name');
-        $last_name    = \Input::get('last_name');
-        $email         = \Input::get('email');
-        $password     = \Input::get('password');
-        $role         = \Input::get('role');
-        $activated     = (\Input::get('activated') == '' ? false : true);
-        
-        // Check if this is logged in user, prevent deactivate
-        if (isset($sid)) {
-            $this_user = Auth::user();
-            if ($this_user->id == $sid && $activated == false) {
-                $errors = new \Illuminate\Support\MessageBag;
+            $path ='admin/users/edit/' . $sid;
+            $user = User::find($sid);
+            
+            // Check if this is logged in user, prevent deactivate
+            if (Auth::user()->id == $sid && $activated == false) {
                 $errors->add(
                     'deactivateError',
                     Lang::get('redminportal::messages.user_error_deactivate_own_account')
                 );
                 return redirect($path)->withErrors($errors)->withInput();
             }
+        } else {
+            $rules['password'] = 'required|confirmed|min:6';
+            $path = 'admin/users/create';
+            $user = new User;
         }
         
-        $user = (isset($sid) ? User::find($sid) : new User);
+        $messages = array(
+            'permission-inherit.regex' => 'The permission inherit format is invalid. Try using the Permission Builder.',
+            'permission-allow.regex' => 'The permission allow format is invalid. Try using the Permission Builder.',
+            'permission-deny.regex' => 'The permission deny format is invalid. Try using the Permission Builder.'
+        );
+
+        $validation = Validator::make(Input::all(), $rules, $messages);
         
+        if ($validation->fails()) {
+            return redirect($path)->withErrors($validation)->withInput();
+        }
+        
+        // If user can't be created or found
         if ($user == null) {
-            $errors = new \Illuminate\Support\MessageBag;
             $errors->add(
                 'createError',
                 Lang::get('redminportal::messages.user_error_create_unknown')
@@ -136,17 +162,26 @@ class UserController extends Controller
             return redirect('/admin/users')->withErrors($errors);
         }
         
+        $permissions = $this->populatePermission(
+            Input::get('permission-inherit'),
+            Input::get('permission-allow'),
+            Input::get('permission-deny')
+        );
+        
         // Save or Update
-        $user->email = $email;
+        $user->email = Input::get('email');
+        
+        $password = Input::get('password');
         if ($password != '') {
-            $user->password = \Hash::make($password);
+            $user->password = Hash::make($password);
         }
-        $user->first_name = $first_name;
-        $user->last_name = $last_name;
+        
+        $user->first_name = Input::get('first_name');
+        $user->last_name = Input::get('last_name');
         $user->activated = $activated;
+        $user->permissions = json_encode($permissions);
         
         if (! $user->save()) {
-            $errors = new \Illuminate\Support\MessageBag;
             $errors->add(
                 'saveError',
                 Lang::get('redminportal::messages.user_error_update_unknown')
@@ -155,11 +190,8 @@ class UserController extends Controller
         }
         
         // Assign group(s) to user
-        $add_group_success = $user->addGroup($role);
-        
         // Return error message if group has error
-        if (! $add_group_success) {
-            $errors = new \Illuminate\Support\MessageBag;
+        if (! $user->addGroup(Input::get('role'))) {
             $errors->add(
                 'groupError',
                 Lang::get('redminportal::messages.user_error_group_not_found')
@@ -174,7 +206,7 @@ class UserController extends Controller
     {
         $this_user = Auth::user();
         if ($this_user->id == $sid) {
-            $errors = new \Illuminate\Support\MessageBag;
+            $errors = new MessageBag;
             $errors->add(
                 'deleteError',
                 Lang::get('redminportal::messages.user_error_delete_own_account')
@@ -185,7 +217,7 @@ class UserController extends Controller
         $user = User::find($sid);
         
         if ($user == null) {
-            $errors = new \Illuminate\Support\MessageBag;
+            $errors = new MessageBag;
             $errors->add(
                 'deleteError',
                 Lang::get('redminportal::messages.user_error_user_not_found')
@@ -204,7 +236,7 @@ class UserController extends Controller
         $user = User::find($sid);
         
         if ($user == null) {
-            $errors = new \Illuminate\Support\MessageBag;
+            $errors = new MessageBag;
             $errors->add(
                 'editError',
                 Lang::get('redminportal::messages.user_error_user_not_found')
@@ -223,7 +255,7 @@ class UserController extends Controller
     {
         $this_user = Auth::user();
         if ($this_user->id == $sid) {
-            $errors = new \Illuminate\Support\MessageBag;
+            $errors = new MessageBag;
             $errors->add(
                 'editError',
                 Lang::get('redminportal::messages.user_error_deactivate_own_account')
@@ -234,7 +266,7 @@ class UserController extends Controller
         $user = User::find($sid);
         
         if ($user == null) {
-            $errors = new \Illuminate\Support\MessageBag;
+            $errors = new MessageBag;
             $errors->add(
                 'editError',
                 Lang::get('redminportal::messages.user_error_user_not_found')
@@ -257,13 +289,13 @@ class UserController extends Controller
             'search' => 'required|regex:' . $pattern
         );
 
-        $validation = \Validator::make(\Input::all(), $rules);
+        $validation = Validator::make(Input::all(), $rules);
 
         if ($validation->fails()) {
             return redirect('admin/users')->withErrors($validation)->withInput();
         }
         
-        $search = trim(\Input::get('search'));
+        $search = trim(Input::get('search'));
         
         // Check special characters
         $errors = $this->checkSpecialChar($search);
@@ -293,7 +325,7 @@ class UserController extends Controller
             'search' => $search
         );
 
-        $validation = \Validator::make($inputs, $rules);
+        $validation = Validator::make($inputs, $rules);
 
         if ($validation->fails()) {
             return redirect('admin/users')->withErrors($validation)->with('search', $search);
@@ -336,7 +368,7 @@ class UserController extends Controller
             'search' => $search
         );
 
-        $validation = \Validator::make($inputs, $rules);
+        $validation = Validator::make($inputs, $rules);
 
         if ($validation->fails()) {
             return redirect('admin/users')->withErrors($validation)->with('search', $search);
@@ -376,7 +408,7 @@ class UserController extends Controller
     private function checkSpecialChar($search)
     {
         if (preg_match("/[%$&*]/i", $search)) {
-            $errors = new \Illuminate\Support\MessageBag;
+            $errors = new MessageBag;
             $errors->add(
                 'stringError',
                 Lang::get('redminportal::messages.error_remove_special_characters')