+++ src/App/Models/User.php
@@ -63,8 +63,22 @@ class User extends Model implements AuthenticatableContract, CanResetPasswordCon
         return $this->hasMany('Redooor\Redminportal\App\Models\Order');
     }
     
+    /**
+     * Converts permission json string to array
+     * Returns an empty array if string is null
+     **/
+    public function permissions()
+    {
+        if ($this->permissions) {
+            return json_decode($this->permissions);
+        }
+        
+        return array();
+    }
+    
     public function delete()
     {
+        // Detach all groups from this user
         $this->groups()->detach();
         
         return parent::delete();