+++ CHANGELOG.md
@@ -61,7 +61,8 @@ This update introduces some new tables and lots of changes.
     - /api/email: Unused, returns an empty JSON list
     - /api/email/all: Get JSON list of emails
     - Check [list of Private API on github Wiki page](https://github.com/redooor/redminportal/wiki/Private-API)
-15. UI: Group permission management, able to edit group permission for any route. (issue 69)
+15. UI: Group permission management, able to edit group permission for any route. (issue #69)
+16. UI: User permission management, able to edit user permission for any route. (issue #69)
 
 ### Enhancements:
 1. Category model missing relationships with other models. (issue #140)